refactor : chat과 관련된 type 분리 및 리펙토링
