--- src/pages/ChatroomDetail/chatroomDetailData.ts
@@ -1,119 +1,0 @@
-export const testChatroomDetailShare = {
-  id: 1,
-  thumbnailImageUrl: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-  title:
-    '채팅 상세 테스트 채팅 상세 테스트 채팅 상세 테스트 채팅 상세 테스트 채팅 상세 테스트 채팅 상세 테스트 채팅 상세 테스트 채팅 상세 테스트 채팅 상세 테스트 채팅 상세 테스트',
-  price: 10000,
-  originalPrice: 40000,
-  currentRecruitment: 3,
-  finalRecruitment: 4,
-};
-
-export const testChatroomDetailChats = [
-  {
-    contents:
-      '안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 ',
-    writer: '닉네임1',
-    writerThumbnailImageUrl: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-    writtenDateTime: '2022-12-30 14:00',
-    writtenByMe: false,
-  },
-  {
-    contents:
-      '안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 ',
-    writer: '닉네임2',
-    writerThumbnailImageUrl: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-    writtenDateTime: '2022-12-30 14:01',
-    writtenByMe: true,
-  },
-  {
-    contents:
-      '안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 ',
-    writer: '닉네임1',
-    writerThumbnailImageUrl: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-    writtenDateTime: '2022-12-30 14:00',
-    writtenByMe: false,
-  },
-  {
-    contents:
-      '안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 ',
-    writer: '닉네임2',
-    writerThumbnailImageUrl: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-    writtenDateTime: '2022-12-30 14:01',
-    writtenByMe: true,
-  },
-  {
-    contents:
-      '안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 ',
-    writer: '닉네임1',
-    writerThumbnailImageUrl: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-    writtenDateTime: '2022-12-30 14:00',
-    writtenByMe: false,
-  },
-  {
-    contents:
-      '안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 ',
-    writer: '닉네임2',
-    writerThumbnailImageUrl: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-    writtenDateTime: '2022-12-30 14:01',
-    writtenByMe: true,
-  },
-  {
-    contents:
-      '안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 ',
-    writer: '닉네임1',
-    writerThumbnailImageUrl: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-    writtenDateTime: '2022-12-30 14:00',
-    writtenByMe: false,
-  },
-  {
-    contents:
-      '안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 ',
-    writer: '닉네임2',
-    writerThumbnailImageUrl: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-    writtenDateTime: '2022-12-30 14:01',
-    writtenByMe: true,
-  },
-  {
-    contents:
-      '안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 ',
-    writer: '닉네임1',
-    writerThumbnailImageUrl: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-    writtenDateTime: '2022-12-30 14:00',
-    writtenByMe: false,
-  },
-  {
-    contents:
-      '안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 ',
-    writer: '닉네임2',
-    writerThumbnailImageUrl: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-    writtenDateTime: '2022-12-30 14:01',
-    writtenByMe: true,
-  },
-  {
-    contents:
-      '안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 안녕하세요 ',
-    writer: '닉네임1',
-    writerThumbnailImageUrl: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-    writtenDateTime: '2022-12-30 14:00',
-    writtenByMe: false,
-  },
-  {
-    contents: '마지막 부분',
-    writer: '닉네임2',
-    writerThumbnailImageUrl: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-    writtenDateTime: '2022-12-30 14:01',
-    writtenByMe: true,
-  },
-];
-
-export const testChatroomDetailData = {
-  share: testChatroomDetailShare,
-  chats: testChatroomDetailChats,
-};
-
-export type TestChatroomDetailShareType = typeof testChatroomDetailShare;
-
-export type TestChatroomDetailChatsType = typeof testChatroomDetailChats;
-
-export type TestChatroomDetailDataType = typeof testChatroomDetailData;
