--- src/pages/Chatrooms/chatting.ts
@@ -1,141 +1,0 @@
-export const testChatroomsInfo = [
-  {
-    id: 0,
-    writers: [
-      {
-        id: 0,
-        writer: '테스트1',
-        img: 'https://avatars.githubusercontent.com/u/92966772?s=64&v=4',
-      },
-    ],
-    content: '테스트 콘텐츠',
-    location: '역삼동',
-    time: '2022-08-17 00:00',
-  },
-  {
-    id: 1,
-    writers: [
-      {
-        id: 0,
-        writer: '테스트1',
-        img: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-      },
-      {
-        id: 1,
-        writer: '테스트2',
-        img: 'https://avatars.githubusercontent.com/u/109581256?s=200&v=4',
-      },
-    ],
-    content:
-      '테스트 콘텐츠 테스트 콘텐츠 테스트 콘텐츠 테스트 콘텐츠 테스트 콘텐츠 테스트 콘텐츠 테스트 콘텐츠 테스트 콘텐츠 테스트 콘텐츠 테스트 콘텐츠 테스트 콘텐츠',
-    location: '역삼동',
-    time: '2022-08-17 00:00',
-  },
-  {
-    id: 2,
-    writers: [
-      {
-        id: 0,
-        writer: '테스트1',
-        img: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-      },
-      {
-        id: 1,
-        writer: '테스트2',
-        img: 'https://avatars.githubusercontent.com/u/109581256?s=200&v=4',
-      },
-      {
-        id: 2,
-        writer: '테스트3',
-        img: 'https://avatars.githubusercontent.com/u/93566353?s=64&v=4',
-      },
-    ],
-    content: '테스트 콘텐츠',
-    location: '역삼동',
-    time: '2022-08-17 00:00',
-  },
-  {
-    id: 3,
-    writers: [
-      {
-        id: 0,
-        writer: '테스트1',
-        img: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-      },
-      {
-        id: 1,
-        writer: '테스트2',
-        img: 'https://avatars.githubusercontent.com/u/92966772?s=64&v=4',
-      },
-      {
-        id: 2,
-        writer: '테스트3',
-        img: 'https://avatars.githubusercontent.com/u/93566353?s=64&v=4',
-      },
-      {
-        id: 3,
-        writer: '테스트4',
-        img: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-      },
-    ],
-    content: '테스트 콘텐츠',
-    location: '역삼동',
-    time: '2022-08-17 00:00',
-  },
-  {
-    id: 4,
-    writers: [
-      {
-        id: 0,
-        writer: '테스트1',
-        img: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-      },
-      {
-        id: 1,
-        writer: '테스트1',
-        img: 'https://avatars.githubusercontent.com/u/92699009?v=4',
-      },
-      {
-        id: 2,
-        writer: '테스트1',
-        img: 'https://avatars.githubusercontent.com/u/109581256?s=200&v=4',
-      },
-      {
-        id: 3,
-        writer: '테스트1',
-        img: 'https://avatars.githubusercontent.com/u/93566353?s=64&v=4',
-      },
-    ],
-    content: '테스트 콘텐츠',
-    location: '역삼동',
-    time: '2022-08-17 00:00',
-  },
-  {
-    id: 5,
-    writers: [
-      {
-        id: 0,
-        writer: '테스트1',
-        img: 'https://avatars.githubusercontent.com/u/67730358?v=4',
-      },
-      {
-        id: 1,
-        writer: '테스트1',
-        img: 'https://avatars.githubusercontent.com/u/109581256?s=200&v=4',
-      },
-      {
-        id: 2,
-        writer: '테스트1',
-        img: 'https://avatars.githubusercontent.com/u/92699009?v=4',
-      },
-      {
-        id: 3,
-        writer: '테스트1',
-        img: 'https://avatars.githubusercontent.com/u/93566353?s=64&v=4',
-      },
-    ],
-    content: '테스트 콘텐츠',
-    location: '역삼동',
-    time: '2022-08-17 00:00',
-  },
-];
